Terrace tool: Dialog for house numbers.
Terrace dialog to choose the number of houses, or alternatively the
house numbers from which the number of houses is derived. Can specify
numbers as ranges separated by semicolons, and can select to use only
odd or only even numbers for those ranges. Can also select in addition
to the area an extra node which is used to indicate the side of the
buildling that house numbers start.

Also improved use of QInputDialog (subdivide tool) so it uses exec()
instead of show().