Fix removal of multiply included features.
When a node was removed from a way, or a feature from a relation, the
parent unparented itself from the child regardless of whether the child
was included multiple times. This is fixed by removing and then only
unparenting if the child is no longer found.

This shows itself when a closed way is undone (the last node which is
equal to the first node is removed), and then the way is deleted.
Way::deleteChildren sees that the node doesn't have exactly one parent
(it has zero parents) and doesn't delete it.