Fix several issues with features dock.
If a highlighted item is deleted it is now removed from the features
list when the list is updated instead of persisting until it is
unhiglighted (FeaturesDock::clearItems).

Features dock is invalidated (lists cleared) prior to deleting the
document. This prevents seg faults if features are highlighted when a
document is closed.

MainWindow now connects theDocument::historyChanged to content_changed()
so that the features dock list gets updated when objects do.