ADD : Join areas tool
Tool for joining together areas which share edges. Tied into gui and
includes a basic icon (which I'm sure could use improvement or
replacement).

Areas must actually be sharing a sequence of nodes for them to be
joined.

The algorithm handles reasonable areas which are thrown at it except
that it does not produce multipolygon relations when an enclosed area is
formed. Instead it ends up with a single area way that cuts into the
middle and back out again.