ADD : Tool to align edges to regular angles.
This is a generalisation of what may be called orthogonalise to make all
angles right angles. The number of regular axes (evenly spaced in the
current projection) is specified (between 3 and 16) and the edges in the
ways (road or area) are made parallel to one of the axes. This is handy
for correcting hand drawn buildings. A rough icon has been drawn which
is welcome to be replaced.

It finds the angle of the first axis and assigns axes using a variation
of k-means clustering of the edge angles, but where the clusters are
evenly separated and the mean is weighted towards longer edges. The
nodes are repositioned on the intersections of the assigned axes through
their respective edge's midpoints. There are several cases where an
intersection cannot be used:
- End nodes of an unclosed way (only one neighbouring edge). The node is
  moved to the closest point on the axis through the edge midpoint.
- Adjacent edges assigned the same axis are handled by detecting them
  first and projecting the edge midpoints onto the axis through the mean
  midpoint, then instead of intersecting we move the node to the closest
  point on the axis through one of the midpoints.
- Adjacent edges assigned opposite parallel axes (when number of axes is
  even) are not handled at the moment and the affected nodes are left
  unmoved.

Nodes shared between multiple selected ways usually prevent the result
from being fully axis aligned.

Also added myself to AUTHORS file.