Axis align: default to guess of number of axes.
Instead of defaulting to 4 axes, now default to a guess of the number of
important axes in the ways, and still allow the user to change it. This
guess is worked out by trying to align the ways to each number of axes
between 3 and 16, and choosing the one with the lowest variance between
the original angles and the aligned angles. The variances are weighted
to favour a smaller number of axes.