Axis align: Display error when unable to complete.
When there's a very sharp corner in the shape and aligning to an even
number of axes, it can't resolve it. This is because the two edges get
classified to opposite axes which cannot be intersected to find the node
position. It's not obvious what should be done in this case.

A message box is now displayed asking the user to resolve the corner
manually and try again.