Axis align: Iterate to handle shared nodes better.
The algorithm didn't handle nodes that were shared between ways or in a
way multiple times. It just made the shape "more" aligned but not
precisely accurate.

The main correction process is now attempted multiple times (20) if
nodes are shared, until the nodes stop moving. In the simple cases this
will converge quickly onto a correct solution. In other cases it will
not (it depends on how the ways are joined, and the order of selection).

If after the iterations the result has still not converged
satisfactorily the user is asked (with a discardable message box)
whether they wish to continue with an innaccurate alignment.