Axis align: Handle shared nodes.
Handle ways that share nodes or are self intersecting. Previously shared
nodes got moved back and forth and would be unlikely to converge on a
solution. Now, if nodes are shared, nodes are only moved half the
distance towards their intersection points, allowing different ways to
compete with one another over a node position. The iteration continues
while the amount the nodes are moving is decreasing, until the movement
is sufficiently small. If the algorithm doesn't converge then it's
likely the ways cannot be aligned to the classified axes.

Since failure is now likely to be due to an unalignable selection, the
box asking the user if they wish to accept the inaccurate result is
replaced with a straight forward error message.