FIX : Fix scale bar to take latitude into account.
Scale bar doesn't currently take latitude into account, so if you draw a
way between the two ends of the scale anywhere other than the equator it
shows a smaller distance than the scale.

It now recalculates MapView::p->PixelPerM in viewportRecalc which gets
called when panning as well as zooming (changing the latitude changes
the scale). For the longitude/latitude projection, it does this using
lonAnglePerM(lat) instead of LAT_ANG_PER_M, and for other projections it
finds the distance between the coordinates of the mid left and mid right
of the screen.