FIX: Fixed random crashes while changing styles.
When style is changed, every Featature needs to be updated with a new
set of painters. This was not done properly, as the Painters may have
been deleted, while the drawing code was running and using them. This
causes occasional crashes after style is changed, and a big area is
being drawn (and the code does not finish drawing before the style is
changed).

I added a lock for the painters, but it is a temporary fix. It will most
probably change to a document-wide lock, since the features should not
be changed at all while being drawn. But this requires much more
care.