Fixed the printing.
The code tried to copy all the tiles before they were rendered. I fixed it with
a busy wait, but it will require more fixing, especially with big resolutions,
as the rendering won't be quick.

The qualisty still sucks, since the QPrinter uses very low resolution, and
someting horribly goes wrong when I change it to 600 or higher.