Fixed the relation viewer to filter correctly.
The previous function interfered with a feature that always leaves selected
feature in the list.

The feature is now highlighted as before, but if it does not matched the filter,
it's not in the list.

This was problematic, as the feature filter was not that useful for checking
if an object is in a relation.

I replaced it by the following: in the features dock, every highlighted feature
is automatically selected. I think this maps to the previous intent.