Cleaned up translation lookup code.
The old code worked fine, but it was hard to know what's going on with all the
ifs and #ifdefs. The new code is a bit simpler and not really equivalent, but I
think it will work the same in practice.

The only problem would be if we supply wrong TRANSDIR_SYSTEM, that points to a
translations for another qt version. But that's bad idea anyway.